refactor: rearrange LedgerFacade methods into conceptual groupings
