+++ src/services/ledger/facade/index.ts
@@ -1,24 +1,7 @@
-import { paymentAmountFromNumber } from "@domain/shared"
-
-import { MainBook } from "../books"
-import { toLedgerAccountId } from "../domain"
-import { UnknownLedgerError } from "../domain/errors"
-
-export * from "./get-txns"
-export * from "./record-txns"
+export * from "./intraledger"
+export * from "./offchain-receive"
+export * from "./offchain-send"
+export * from "./onchain-receive"
+export * from "./onchain-send"
 export * from "./tx-metadata"
-export * from "./update-txns"
-
-export const getLedgerAccountBalanceForWalletId = async <T extends WalletCurrency>({
-  id: walletId,
-  currency,
-}: WalletDescriptor<T>): Promise<PaymentAmount<T> | LedgerError> => {
-  try {
-    const { balance } = await MainBook.balance({
-      account: toLedgerAccountId(walletId),
-    })
-    return paymentAmountFromNumber({ amount: balance, currency })
-  } catch (err) {
-    return new UnknownLedgerError(err)
-  }
-}
+export * from "./wallet"
